Phase-3 Removal vavr dependency (#1156)
