+++ resilience4j-cache/build.gradle
@@ -2,6 +2,7 @@ dependencies {
     api project(':resilience4j-core')
     api (libraries.jcache)
     testImplementation project(':resilience4j-rxjava2')
+    testImplementation project(':resilience4j-vavr')
     testImplementation(libraries.rxjava2)
 }
 ext.moduleName = 'io.github.resilience4j.cache'
