+++ resilience4j-cache/src/test/java/io/github/resilience4j/cache/CacheEventPublisherTest.java
@@ -56,7 +56,7 @@ public class CacheEventPublisherTest {
         Cache<String, String> cacheContext = Cache.of(cache);
         cacheContext.getEventPublisher().onCacheHit(
             event -> logger.info(event.getEventType().toString()));
-        CheckedFunction1<String, String> cachedFunction = Cache
+        CheckedFunction1<String, String> cachedFunction = VavrCache
             .decorateCheckedSupplier(cacheContext, () -> "Hello world");
 
         String value = cachedFunction.apply("testKey");
@@ -71,7 +71,7 @@ public class CacheEventPublisherTest {
         Cache<String, String> cacheContext = Cache.of(cache);
         cacheContext.getEventPublisher().onCacheMiss(
             event -> logger.info(event.getEventType().toString()));
-        CheckedFunction1<String, String> cachedFunction = Cache
+        CheckedFunction1<String, String> cachedFunction = VavrCache
             .decorateCheckedSupplier(cacheContext, () -> "Hello world");
 
         String value = cachedFunction.apply("testKey");
@@ -86,7 +86,7 @@ public class CacheEventPublisherTest {
         Cache<String, String> cacheContext = Cache.of(cache);
         cacheContext.getEventPublisher().onError(
             event -> logger.info(event.getEventType().toString()));
-        CheckedFunction1<String, String> cachedFunction = Cache
+        CheckedFunction1<String, String> cachedFunction = VavrCache
             .decorateCheckedSupplier(cacheContext, () -> "Hello world");
         String value = cachedFunction.apply("testKey");
