+++ resilience4j-circularbuffer/src/main/java/io/github/resilience4j/circularbuffer/CircularFifoBuffer.java
@@ -18,8 +18,9 @@
  */
 package io.github.resilience4j.circularbuffer;
 
-import io.vavr.collection.List;
-import io.vavr.control.Option;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Stream;
 
 /**
  * A {@link CircularFifoBuffer} is a first in first out buffer with a fixed size that replaces its
@@ -57,6 +58,13 @@ public interface CircularFifoBuffer<T> {
     List<T> toList();
 
     /**
+     * Returns a stream of the elements in this {@link CircularFifoBuffer}.
+     *
+     * @return a stream of the elements in this {@link CircularFifoBuffer}
+     */
+    Stream<T> toStream();
+
+    /**
      * Adds element to the {@link CircularFifoBuffer} and overwrites the oldest element when {@link
      * CircularFifoBuffer#isFull}.
      *
@@ -66,10 +74,10 @@ public interface CircularFifoBuffer<T> {
     void add(T element);
 
     /**
-     * Retrieves and removes the head of this queue, or returns {@link Option.None} if this queue is
+     * Retrieves and removes the head of this queue, or returns {@link Optional#empty()} if this queue is
      * empty.
      *
-     * @return the head of this queue, or {@link Option.None} if this queue is empty
+     * @return the head of this queue, or {@link Optional#empty()} empty} if this queue is empty
      */
-    Option<T> take();
+    Optional<T> take();
 }
