+++ resilience4j-circularbuffer/src/test/java/io/github/resilience4j/circularbuffer/CircularFifoBufferTest.java
@@ -19,11 +19,11 @@
 package io.github.resilience4j.circularbuffer;
 
 
-import io.vavr.collection.List;
 import org.junit.Test;
 
 import java.io.IOException;
 import java.net.UnknownHostException;
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -55,6 +55,10 @@ public class CircularFifoBufferTest {
         assertThat(bufferedExceptions.get(1)).isInstanceOf(IOException.class);
         assertThat(bufferedExceptions.get(2)).isInstanceOf(IllegalStateException.class);
         assertThat(bufferedExceptions.get(3)).isInstanceOf(UnknownHostException.class);
+
+        assertThat(exceptionBuffer.toStream()).hasSize(4);
+        assertThat(exceptionBuffer.toStream()).hasOnlyElementsOfTypes(IllegalArgumentException.class,
+            IOException.class, IllegalStateException.class, UnknownHostException.class);
 
         // The size must still be 4, because the CircularFifoBuffer capacity is 4
         exceptionBuffer.add(new IOException("bla bla"));
