+++ resilience4j-consumer/src/main/java/io/github/resilience4j/consumer/CircularEventConsumer.java
@@ -21,7 +21,9 @@ package io.github.resilience4j.consumer;
 import io.github.resilience4j.circularbuffer.CircularFifoBuffer;
 import io.github.resilience4j.circularbuffer.ConcurrentCircularFifoBuffer;
 import io.github.resilience4j.core.EventConsumer;
-import io.vavr.collection.List;
+
+import java.util.List;
+import java.util.stream.Stream;
 
 /**
  * A consumer which stores CircuitBreakerEvents in a circular buffer with a fixed capacity.
@@ -52,5 +54,14 @@ public class CircularEventConsumer<T> implements EventConsumer<T> {
      */
     public List<T> getBufferedEvents() {
         return eventCircularFifoBuffer.toList();
+    }
+
+    /**
+     * Returns a stream containing all of the buffered events.
+     *
+     * @return a stream containing all of the buffered events.
+     */
+    public Stream<T> getBufferedEventsStream() {
+        return eventCircularFifoBuffer.toStream();
     }
 }
