+++ resilience4j-metrics/src/main/java/io/github/resilience4j/metrics/CircuitBreakerMetrics.java
@@ -24,8 +24,8 @@ import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.MetricSet;
 import io.github.resilience4j.circuitbreaker.CircuitBreaker;
 import io.github.resilience4j.circuitbreaker.CircuitBreakerRegistry;
-import io.vavr.collection.Array;
 
+import java.util.List;
 import java.util.Map;
 
 import static com.codahale.metrics.MetricRegistry.name;
@@ -155,7 +155,7 @@ public class CircuitBreakerMetrics implements MetricSet {
      * @param circuitBreaker the circuit breaker
      */
     public static CircuitBreakerMetrics ofCircuitBreaker(CircuitBreaker circuitBreaker) {
-        return new CircuitBreakerMetrics(Array.of(circuitBreaker));
+        return new CircuitBreakerMetrics(List.of(circuitBreaker));
     }
 
     @Override
