+++ resilience4j-metrics/src/main/java/io/github/resilience4j/metrics/RetryMetrics.java
@@ -6,8 +6,8 @@ import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.MetricSet;
 import io.github.resilience4j.retry.Retry;
 import io.github.resilience4j.retry.RetryRegistry;
-import io.vavr.collection.Array;
 
+import java.util.List;
 import java.util.Map;
 
 import static com.codahale.metrics.MetricRegistry.name;
@@ -72,7 +72,7 @@ public class RetryMetrics implements MetricSet {
     }
 
     public static RetryMetrics ofRetry(Retry retry) {
-        return new RetryMetrics(Array.of(retry));
+        return new RetryMetrics(List.of(retry));
     }
 
     @Override
