+++ resilience4j-metrics/src/main/java/io/github/resilience4j/metrics/TimeLimiterMetrics.java
@@ -24,8 +24,8 @@ import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.MetricSet;
 import io.github.resilience4j.timelimiter.TimeLimiter;
 import io.github.resilience4j.timelimiter.TimeLimiterRegistry;
-import io.vavr.collection.Array;
 
+import java.util.List;
 import java.util.Map;
 
 import static com.codahale.metrics.MetricRegistry.name;
@@ -139,7 +139,7 @@ public class TimeLimiterMetrics implements MetricSet {
      * @param timeLimiter the time limiter
      */
     public static TimeLimiterMetrics ofTimeLimiter(TimeLimiter timeLimiter) {
-        return new TimeLimiterMetrics(Array.of(timeLimiter));
+        return new TimeLimiterMetrics(List.of(timeLimiter));
     }
 
     @Override
