+++ resilience4j-ratpack/src/main/java/io/github/resilience4j/ratpack/circuitbreaker/monitoring/endpoint/CircuitBreakerChain.java
@@ -108,10 +108,11 @@ public class CircuitBreakerChain implements Action<Chain> {
                     CircuitBreakerEventsEndpointResponse response = new CircuitBreakerEventsEndpointResponse(
                         eventConsumerRegistry
                             .getAllEventConsumer()
-                            .flatMap(CircularEventConsumer::getBufferedEvents)
+                            .stream()
+                            .flatMap(CircularEventConsumer::getBufferedEventsStream)
                             .sorted(Comparator.comparing(CircuitBreakerEvent::getCreationTime))
                             .map(CircuitBreakerEventDTOFactory::createCircuitBreakerEventDTO)
-                            .toJavaList());
+                            .collect(Collectors.toList()));
                     d.success(response);
                 }).then(r -> ctx.render(Jackson.json(r)))
             );
@@ -130,21 +131,22 @@ public class CircuitBreakerChain implements Action<Chain> {
                 ctx.render(events);
             });
             chain1.get("hystrixStream/events", ctx -> {
-                Seq<Flux<CircuitBreakerEvent>> eventStreams = circuitBreakerRegistry
-                    .getAllCircuitBreakers().map(circuitBreaker -> ReactorAdapter
-                        .toFlux(circuitBreaker.getEventPublisher()));
+                Flux<CircuitBreakerEvent> eventStreams = Flux.fromIterable(circuitBreakerRegistry.getAllCircuitBreakers())
+                    .flatMap(circuitBreaker -> ReactorAdapter.toFlux(circuitBreaker.getEventPublisher()));
                 Function<CircuitBreakerEvent, String> data = c -> Jackson.getObjectWriter(chain1.getRegistry())
                     .writeValueAsString(
                         circuitBreakerRegistry.getAllCircuitBreakers()
+                            .stream()
                             .filter(cb -> cb.getName().equals(c.getCircuitBreakerName()))
                             .map(cb -> new CircuitBreakerHystrixStreamEventsDTO(c,
                                 cb.getState(),
                                 cb.getMetrics(),
                                 cb.getCircuitBreakerConfig()
                             ))
+                            .collect(Collectors.toList())
                     );
                 ServerSentEvents events = ServerSentEvents
-                    .serverSentEvents(Flux.merge(eventStreams),
+                    .serverSentEvents(eventStreams,
                         e -> e.id(CircuitBreakerEvent::getCircuitBreakerName)
                             .event(c -> c.getEventType().name()).data(data));
                 ctx.render(events);
@@ -155,9 +157,9 @@ public class CircuitBreakerChain implements Action<Chain> {
                         CircuitBreakerEventsEndpointResponse response = new CircuitBreakerEventsEndpointResponse(
                             eventConsumerRegistry
                                 .getEventConsumer(circuitBreakerName)
-                                .getBufferedEvents()
+                                .getBufferedEventsStream()
                                 .map(CircuitBreakerEventDTOFactory::createCircuitBreakerEventDTO)
-                                .toJavaList());
+                                .collect(Collectors.toList()));
                         d.success(response);
                     }).then(r -> ctx.render(Jackson.json(r)));
                 }
@@ -181,8 +183,10 @@ public class CircuitBreakerChain implements Action<Chain> {
             chain1.get("hystrixStream/events/:name", ctx -> {
                 String circuitBreakerName = ctx.getPathTokens().get("name");
                 CircuitBreaker circuitBreaker = circuitBreakerRegistry.getAllCircuitBreakers()
-                    .find(cb -> cb.getName().equals(circuitBreakerName))
-                    .getOrElseThrow(() -> new IllegalArgumentException(String
+                    .stream()
+                    .filter(cb -> cb.getName().equals(circuitBreakerName))
+                    .findAny()
+                    .orElseThrow(() -> new IllegalArgumentException(String
                         .format("circuit breaker with name %s not found", circuitBreakerName)));
                 Function<CircuitBreakerEvent, String> data = c -> Jackson.getObjectWriter(chain1.getRegistry())
                     .writeValueAsString(
@@ -205,11 +209,11 @@ public class CircuitBreakerChain implements Action<Chain> {
                         CircuitBreakerEventsEndpointResponse response = new CircuitBreakerEventsEndpointResponse(
                             eventConsumerRegistry
                                 .getEventConsumer(circuitBreakerName)
-                                .getBufferedEvents()
+                                .getBufferedEventsStream()
                                 .filter(event -> event.getEventType() == CircuitBreakerEvent.Type
                                     .valueOf(eventType.toUpperCase()))
                                 .map(CircuitBreakerEventDTOFactory::createCircuitBreakerEventDTO)
-                                .toJavaList());
+                                .collect(Collectors.toList()));
                         d.success(response);
                     }).then(r -> ctx.render(Jackson.json(r)));
                 }
@@ -237,9 +241,10 @@ public class CircuitBreakerChain implements Action<Chain> {
             chain1.get("hystrixStream/events/:name/:type", ctx -> {
                 String circuitBreakerName = ctx.getPathTokens().get("name");
                 String eventType = ctx.getPathTokens().get("type");
-                CircuitBreaker circuitBreaker = circuitBreakerRegistry.getAllCircuitBreakers()
-                    .find(cb -> cb.getName().equals(circuitBreakerName))
-                    .getOrElseThrow(() -> new IllegalArgumentException(String
+                CircuitBreaker circuitBreaker = circuitBreakerRegistry.getAllCircuitBreakers().stream()
+                    .filter(cb -> cb.getName().equals(circuitBreakerName))
+                    .findAny()
+                    .orElseThrow(() -> new IllegalArgumentException(String
                         .format("circuit breaker with name %s not found", circuitBreakerName)));
                 Flux<CircuitBreakerEvent> eventStream = ReactorAdapter
                     .toFlux(circuitBreaker.getEventPublisher())
