+++ resilience4j-ratpack/src/main/java/io/github/resilience4j/ratpack/retry/monitoring/endpoint/RetryChain.java
@@ -37,6 +37,7 @@ import reactor.core.publisher.Flux;
 import javax.inject.Inject;
 import java.util.Comparator;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Provides event and stream event endpoints for circuitbreaker events.
@@ -60,10 +61,11 @@ public class RetryChain implements Action<Chain> {
         chain.prefix(prefix, chain1 -> {
             chain1.get("events", ctx ->
                 Promise.<RetryEventsEndpointResponse>async(d -> {
-                    List<RetryEventDTO> eventsList = eventConsumerRegistry.getAllEventConsumer()
-                        .flatMap(CircularEventConsumer::getBufferedEvents)
+                    List<RetryEventDTO> eventsList = eventConsumerRegistry.getAllEventConsumer().stream()
+                        .flatMap(CircularEventConsumer::getBufferedEventsStream)
                         .sorted(Comparator.comparing(RetryEvent::getCreationTime))
-                        .map(RetryEventDTOFactory::createRetryEventDTO).toJavaList();
+                        .map(RetryEventDTOFactory::createRetryEventDTO)
+                        .collect(Collectors.toList());
                     d.success(new RetryEventsEndpointResponse(eventsList));
                 }).then(r -> ctx.render(Jackson.json(r)))
             );
@@ -84,9 +86,10 @@ public class RetryChain implements Action<Chain> {
                     Promise.<RetryEventsEndpointResponse>async(d -> {
                         List<RetryEventDTO> eventsList = eventConsumerRegistry
                             .getEventConsumer(retryName)
-                            .getBufferedEvents()
+                            .getBufferedEventsStream()
                             .sorted(Comparator.comparing(RetryEvent::getCreationTime))
-                            .map(RetryEventDTOFactory::createRetryEventDTO).toJavaList();
+                            .map(RetryEventDTOFactory::createRetryEventDTO)
+                            .collect(Collectors.toList());
                         d.success(new RetryEventsEndpointResponse(eventsList));
                     }).then(r -> ctx.render(Jackson.json(r)));
                 }
@@ -114,11 +117,12 @@ public class RetryChain implements Action<Chain> {
                     Promise.<RetryEventsEndpointResponse>async(d -> {
                         List<RetryEventDTO> eventsList = eventConsumerRegistry
                             .getEventConsumer(retryName)
-                            .getBufferedEvents()
+                            .getBufferedEventsStream()
                             .sorted(Comparator.comparing(RetryEvent::getCreationTime))
                             .filter(event -> event.getEventType() == RetryEvent.Type
                                 .valueOf(eventType.toUpperCase()))
-                            .map(RetryEventDTOFactory::createRetryEventDTO).toJavaList();
+                            .map(RetryEventDTOFactory::createRetryEventDTO)
+                            .collect(Collectors.toList());
                         d.success(new RetryEventsEndpointResponse(eventsList));
                     }).then(r -> ctx.render(Jackson.json(r)));
                 }
