+++ resilience4j-vavr/build.gradle
@@ -1,12 +1,16 @@
 dependencies {
     compile project(':resilience4j-retry')
     compile project(':resilience4j-circuitbreaker')
+    compile project(':resilience4j-cache')
     compile project(':resilience4j-bulkhead')
     compile project(':resilience4j-ratelimiter')
+    compile project(':resilience4j-metrics')
+    compile project(':resilience4j-prometheus')
     compile ( libraries.vavr )
     testCompile project(':resilience4j-test')
     testCompile project(':resilience4j-rxjava2')
     testCompile(libraries.rxjava2)
+    testCompile(libraries.metrics)
 }
 
 ext.moduleName = 'io.github.resilience4j.vavr'
