+++ resilience4j-vavr/src/main/java/io/github/resilience4j/core/CheckFunctionUtils.java
